+++ src/core/system/render/light.cpp
@@ -1,0 +1,4 @@
+#include "light.h"
+
+#include "raymath.h"
+#include "rlgl.h"
