+++ src/core/system/render/light.h
@@ -1,0 +1,6 @@
+#pragma once
+#include "../system_base.h"
+
+#include "raylib.h"
+
+
