+++ src/core/system/render/shaders/lights/depth.fs
