+++ src/core/system/render/shaders/raylib/alpha_discard.fs
