+++ src/core/system/render/shaders/raylib/base_lighting_instanced.vs
