+++ src/core/system/render/shaders/raylib/bloom.fs
