+++ src/core/system/render/shaders/raylib/color_mix.fs
