+++ src/core/system/render/shaders/raylib/cross_hatching.fs
