+++ src/core/system/render/shaders/raylib/cross_stitching.fs
