+++ src/core/system/render/shaders/raylib/cubes_panning.fs
