+++ src/core/system/render/shaders/raylib/depth.fs
