+++ src/core/system/render/shaders/raylib/dream_vision.fs
