+++ src/core/system/render/shaders/raylib/fog.fs
