+++ src/core/system/render/shaders/raylib/grayscale.fs
