+++ src/core/system/render/shaders/raylib/julia_set.fs
