+++ src/core/system/render/shaders/raylib/lighting.fs
