+++ src/core/system/render/shaders/raylib/mask.fs
