+++ src/core/system/render/shaders/raylib/outline.fs
