+++ src/core/system/render/shaders/raylib/palette_switch.fs
