+++ src/core/system/render/shaders/raylib/posterization.fs
