+++ src/core/system/render/shaders/raylib/predator.fs
