+++ src/core/system/render/shaders/raylib/sobel.fs
