+++ src/core/system/render/shaders/raylib/spotlight.fs
