+++ src/core/system/render/shaders/raylib/swirl.fs
