+++ src/core/system/render/utils.cpp
