+++ src/core/system/render/utils.h
