+++ src/debug/debugger_render.h
@@ -1,4 +1,4 @@
-#include "core/system/render.h"
+#include "core/system/render/render.h"
 
 template<>
 struct SystemDebugger<RenderSystem>
